feat: fix bug in ObjectToInterfaceString and refactor simulator classes
